+++ CHANGELOG.md
@@ -4,6 +4,13 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: `Button`: Change `lithium` modifier styles.
+- Feature: `Button`: Add `active` prop that switches modifier to `lithium` when true.
+- Feature: `TagButton`: Deprecate `selected` prop, use new `active` instead.
+- Fix: `ButtonGroup`: Fix styles.
+- Fix: `TagButton`: Fix styles.
+- Fix: `CheckableButton`: Fix styles.
+
 ## [6.14.0] - 2022-01-03
 
 Features: